+++ Sources/RInAppMessaging/Models/Responses/CampaignDataModels.swift
@@ -72,34 +72,29 @@ internal struct ButtonBehavior: Codable {
 
 internal struct CustomJson: Codable {
     let pushPrimer: PrimerButton?
+    let clickableImage: ClickableImage?
     
     enum CodingKeys: String, CodingKey {
         case pushPrimer
+        case clickableImage
     }
     
-    init(pushPrimer: PrimerButton? = nil) {
+    init(pushPrimer: PrimerButton? = nil, clickableImage: ClickableImage? = nil) {
         self.pushPrimer = pushPrimer
+        self.clickableImage = clickableImage
     }
     
     init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         pushPrimer = try container.decodeIfPresent(PrimerButton.self, forKey: .pushPrimer)
+        clickableImage = try container.decodeIfPresent(ClickableImage.self, forKey: .clickableImage)
     }
 }
 
 internal struct PrimerButton: Codable {
     let button: Int?
-    
-    enum CodingKeys: String, CodingKey {
-        case button
-    }
-    
-    init(button: Int? = nil) {
-        self.button = button
-    }
+}
 
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        button = try container.decodeIfPresent(Int.self, forKey: .button)
-    }
+internal struct ClickableImage: Codable {
+    let url: String?
 }